refactor: Refactoring package structure
