+++ app/src/main/res/drawable-anydpi/ic_add_white.xml
@@ -10,7 +10,7 @@
       android:translateX="1.0909091"
       android:translateY="1.0909091">
     <path
-        android:fillColor="@android:color/white"
+        android:fillColor="#FFFFFF"
         android:pathData="M19,13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z"/>
   </group>
 </vector>
