+++ app/src/main/res/drawable-anydpi/ic_chevron_right_light_gray.xml
@@ -10,7 +10,7 @@
       android:translateX="1.0909091"
       android:translateY="1.0909091">
     <path
-        android:fillColor="@android:color/white"
+        android:fillColor="#FFFFFF"
         android:pathData="M10,6L8.59,7.41 13.17,12l-4.58,4.59L10,18l6,-6z"/>
   </group>
 </vector>
