+++ app/src/main/res/drawable-anydpi/ic_home.xml
@@ -6,6 +6,6 @@
     android:tint="#000000"
     android:alpha="0.8">
   <path
-      android:fillColor="@android:color/white"
+      android:fillColor="#FFFFFF"
       android:pathData="M12,5.69l5,4.5V18h-2v-6H9v6H7v-7.81l5,-4.5M12,3L2,12h3v8h6v-6h2v6h6v-8h3L12,3z"/>
 </vector>
