+++ app/src/main/res/drawable-anydpi/ic_notification.xml
@@ -6,6 +6,6 @@
     android:tint="#000000"
     android:alpha="0.8">
   <path
-      android:fillColor="@android:color/white"
+      android:fillColor="#FFFFFF"
       android:pathData="M12,22c1.1,0 2,-0.9 2,-2h-4c0,1.1 0.9,2 2,2zM18,16v-5c0,-3.07 -1.63,-5.64 -4.5,-6.32L13.5,4c0,-0.83 -0.67,-1.5 -1.5,-1.5s-1.5,0.67 -1.5,1.5v0.68C7.64,5.36 6,7.92 6,11v5l-2,2v1h16v-1l-2,-2zM16,17L8,17v-6c0,-2.48 1.51,-4.5 4,-4.5s4,2.02 4,4.5v6z"/>
 </vector>
