+++ app/src/main/res/drawable/all_roundcorner_orange.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="@color/orange" />
+    <corners android:radius="20dp" />
+</shape>
