+++ app/src/main/res/layout/activity_addstore.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/default_background_color"
     android:orientation="vertical"
-    tools:context=".navigation.addpost.addstore.input.AddStoreActivity">
+    tools:context=".navigation.addpost.insertdata.AddStoreActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/layout_addstore_appbartop"
