+++ app/src/main/res/layout/activity_changeaddress.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/default_background_color"
     android:orientation="vertical"
-    tools:context=".navigation.addpost.addstore.input.ChangeAddressActivity">
+    tools:context=".navigation.addpost.addressing.ChangeAddressActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
