+++ app/src/main/res/layout/activity_choosebuilding.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".navigation.addpost.addstore.addressing.ChooseBuildingActivity">
+    tools:context=".navigation.addpost.addressing.ChooseBuildingActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/layout_choosebuilding_appbartop"
