+++ app/src/main/res/layout/activity_chooseitemcategory.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/default_background_color"
-    tools:context=".navigation.addpost.additem.input.ChooseItemCategoryActivity">
+    tools:context=".navigation.addpost.insertdata.ChooseItemCategoryActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/layout_chooseitemcategory_appbartop"
