+++ app/src/main/res/layout/activity_searchaddress.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".navigation.addpost.addstore.addressing.SearchAddressActivity">
+    tools:context=".navigation.addpost.addressing.SearchAddressActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/layout_searchaddress_appbartop"
